+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @acpus/core
 
+## 0.7.0-alpha.2
+
+### Patch Changes
+
+- Fix installed workflow typechecking for Acpus authoring facade imports.
+- Updated dependencies
+  - @acpus/expression@0.1.0-alpha.2
+
 ## 0.7.0-alpha.1
 
 ### Patch Changes